+++ app/javascript/packages/rails-i18n-webpack-plugin/spec/fixtures/expected1.es.js
@@ -1,1 +1,1 @@
-!function(){var i,t={"forms.button.submit":"Enviar","item.1":"First","item.2":"Second","item.3":""},o=window._locale_data=window._locale_data||{};for(i in t)o[i]=t[i]}();
+!function(){var i,t={"forms.button.submit":"Enviar","item.1":"First","item.2":"Second","item.3":"","forms.button.reset":"Reiniciar"},o=window._locale_data=window._locale_data||{};for(i in t)o[i]=t[i]}();