+++ app/javascript/packages/rails-i18n-webpack-plugin/spec/fixtures/in1.js
@@ -1,4 +1,5 @@
 import './common.js';
+import './resolved';
 
 const text = t('forms.button.submit');