+++ app/javascript/packages/url/index.ts
@@ -2,17 +2,14 @@
  * Given a URL or a string fragment of search parameters and an object of parameters, returns a
  * new URL or search parameters with the parameters added.
  *
- * @param {string} urlOrParams Original URL or search parameters.
- * @param {Record<string, *>} params Search parameters to add.
+ * @param urlOrParams Original URL or search parameters.
+ * @param params Search parameters to add.
  *
- * @return {string} Modified URL or search parameters.
+ * @return Modified URL or search parameters.
  */
-export function addSearchParams(urlOrParams, params) {
-  /** @type {URL|URLSearchParams} */
-  let parsedURLOrParams;
-
-  /** @type {URLSearchParams} */
-  let searchParams;
+export function addSearchParams(urlOrParams: string, params: Record<string, any>): string {
+  let parsedURLOrParams: URL | URLSearchParams;
+  let searchParams: URLSearchParams;
 
   try {
     parsedURLOrParams = new URL(urlOrParams);