+++ app/javascript/packages/eslint-plugin/CHANGELOG.md
@@ -5,6 +5,10 @@
 - Updated peer dependencies to require ESLint >= 8 ([see migration guide](https://eslint.org/docs/8.0.0/user-guide/migrating-to-8.0.0)).
 - Updated base Airbnb shared configurations, introducing newly-enforced rules (see [`airbnb` changelog](https://github.com/airbnb/javascript/blob/master/packages/eslint-config-airbnb/CHANGELOG.md), [`airbnb-base` changelog](https://github.com/airbnb/javascript/blob/master/packages/eslint-config-airbnb-base/CHANGELOG.md)).
 
+### New features
+
+- Added new TypeScript opt-in automatic configuration.
+
 ### Backwards-compatible changes
 
 - `no-cond-assign` is now configured with the [`except-parens` option](https://eslint.org/docs/rules/no-cond-assign#except-parens).